// types/store-settings.ts
export interface StoreSettings {
  id: string;
  logoUrl: string | null;
  primaryColor: string;
  welcomeMessage: string | null;
  messageSignature: string | null;
  slug: string;
  googleMapsUrl: string | null;
  whatsappNumber: string | null;
  unretrievedDays: number;
  storeAddress: string | null;
  businessHours: string | null;
  mapLatitude: number | null;
  mapLongitude: number | null;
  facebookUrl: string | null;
  instagramUrl: string | null;
  tiktokUrl: string | null;
  twitterUrl: string | null;
  youtubeUrl: string | null;
  createdAt: string;
  updatedAt: string;
  storeId: string;
}